FIX: chart repository abstraction
