+++ app/src/main/java/com/example/gardenbotapp/data/domain/OnboardingRepositoryImpl.kt
@@ -1,0 +1,54 @@
+/*
+ * Copyright (c) 2021. Created by charr0max  -> manuelrg88@gmail.com
+ */
+
+package com.example.gardenbotapp.data.domain
+
+import android.util.Log
+import com.apollographql.apollo.api.Response
+import com.apollographql.apollo.coroutines.await
+import com.apollographql.apollo.coroutines.toFlow
+import com.apollographql.apollo.exception.ApolloException
+import com.example.gardenbotapp.ActivateDeviceMutation
+import com.example.gardenbotapp.NewDeviceSubscription
+import com.example.gardenbotapp.data.remote.Client
+import com.example.gardenbotapp.data.remote.GardenBotRepository
+import kotlinx.coroutines.flow.Flow
+
+class OnboardingRepositoryImpl : GardenBotBaseRepositoryImpl(), OnboardingRepository {
+
+    override suspend fun newDeviceSub(deviceName: String): Flow<Response<NewDeviceSubscription.Data>> {
+        try {
+            return Client.getInstance().subscribe(NewDeviceSubscription(deviceName)).toFlow()
+        } catch (e: ApolloException) {
+            Log.i(GardenBotRepository.TAG, "ERROR: ${e.message}")
+            throw ApolloException("${e.message}")
+        }
+    }
+
+    override suspend fun activateDevice(
+        deviceName: String,
+        userId: String,
+        token: String
+    ): String {
+        if (deviceName.isNotEmpty()) {
+            try {
+                val response =
+                    Client.getInstance(token).mutate(ActivateDeviceMutation(deviceName, userId))
+                        .await()
+                if (response.data?.activateDevice == null || response.hasErrors()) {
+                    Log.i(GardenBotRepository.TAG, "ERROR: ${response.errors?.map { it.message }}")
+                    throw ApolloException("${response.errors?.map { it.message }}")
+                } else {
+                    response.data?.let {
+                        return it.activateDevice.id
+                    }
+                }
+            } catch (e: ApolloException) {
+                Log.i(GardenBotRepository.TAG, "ERROR: ${e.message}")
+                throw ApolloException("${e.message}")
+            }
+        }
+        return ""
+    }
+}
